requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
docopt==0.6.2
idna==3.4
joblib==1.2.0
Js2Py==0.74
nltk==3.7
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pipwin==0.5.2
PyAudio @ file:///C:/Users/Samue/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pyFirmata==1.1.0
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pyserial==3.5
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
tqdm==4.64.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
